requirements.txt
django==1.7
easy-thumbnails
pygithub3
python-social-auth
#mongoengine
requests
requests-cache
gunicorn
psycopg2
django-image-cropping
